fix(update-customer-address-usecase): add the if to check address is null before call deleteById

+++ application/src/main/java/com/kaua/ecommerce/application/usecases/customer/update/address/DefaultUpdateCustomerAddressUseCase.java
@@ -61,7 +61,10 @@ public class DefaultUpdateCustomerAddressUseCase extends UpdateCustomerAddressUs
 
         final var aCustomerWithAddress = aCustomer.changeAddress(aAddress);
         this.customerGateway.update(aCustomerWithAddress);
-        this.addressGateway.deleteById(aOldAddressId);
+
+        if (aOldAddressId != null) {
+            this.addressGateway.deleteById(aOldAddressId);
+        }
 
         return Either.right(UpdateCustomerAddressOutput.from(aCustomerWithAddress));
     }
